+++ changelog.md
@@ -5,6 +5,7 @@
 * Consistently count search attempts
 * Make custom file ending optional for thumbnails in MediaFileUrlParser
 * Fail less hard for misrepresented urls in MediaFileUrlParser
+* Ban dash from hex digits in MediaFileUrlParser
 
 ## v0.0.7
 * Add Referer classifier